Done
Pong is a classic two-player arcade game where players control paddles on opposite ends of the screen and try to hit a ball back and forth between them without missing. This implementation of Pong was created using the Pygame library in Python. Players use the "W" and "S" keys to move the left paddle and the up and down arrow keys to move the right paddle. The game ends when one player scores a predetermined number of points, which is set to 5 in this implementation. The game features simple graphics and sound effects and is a fun way to pass the time with a friend or to practice your Pygame skills.